init refactor v2

+++ kbengine/src/kbengine/wordbuffer.h
@@ -20,7 +20,7 @@ using namespace std;
 
 
 struct BufferEntry {
-    UInt8 keyCode;
+    UInt16 keyCode;
     bool cap;    // capitalize or not
     RoofType roofType = ORIGIN;
     short tone = KeyEvent::Tone0;    // KeyEvent enum tone: 0 = none, 1 = sac, 2 = huyen, 3 = hoi, 4 = nga, 5 = nang
